feat(pages): 각 페이지 도메인 폴더 구조로 정리
- pages/dashboard/, chat/, calendar/, attendance/, drive/, ai-chat/, members/, settings/
- Dashboard 출퇴근 로직 features/attendance/model/useWorkSession으로 분리
- 기존 pages/*.tsx는 re-export 스텁으로 교체 (router 호환 유지)

closes #7
