+++ src/pages/attendance/attendance.css
@@ -1,0 +1,220 @@
+.attendance-page {
+  max-width: 1400px;
+}
+
+.attendance-header {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+  gap: 16px;
+  margin-bottom: 24px;
+}
+
+.attendance-header h1 {
+  font-size: 1.75rem;
+  font-weight: 700;
+  margin: 0;
+}
+
+.header-actions {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  flex-wrap: wrap;
+}
+
+.export-btn {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 10px 18px;
+  border-radius: 10px;
+  background: var(--gradient-purple);
+  color: var(--text-primary);
+  font-weight: 500;
+}
+
+.date-range {
+  padding: 10px 16px;
+  font-size: 0.9rem;
+  color: var(--text-secondary);
+}
+
+.filter-tabs button {
+  padding: 8px 16px;
+  border-radius: 8px;
+  background: rgba(148, 163, 184, 0.1);
+  color: var(--text-secondary);
+  font-size: 0.875rem;
+}
+
+.filter-tabs button.active {
+  background: rgba(124, 58, 237, 0.3);
+  color: var(--accent-purple-light);
+}
+
+.attendance-content {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.two-cards-row {
+  display: grid;
+  grid-template-columns: 1fr 1.5fr;
+  gap: 24px;
+}
+
+.two-cards-row.single {
+  grid-template-columns: 1fr;
+  max-width: 500px;
+}
+
+.set-work-days-section {
+  padding: 24px;
+}
+
+.records-section {
+  padding: 24px;
+}
+
+.records-section h3 {
+  margin-bottom: 16px;
+  font-size: 1rem;
+}
+
+.records-table-wrap {
+  overflow-x: auto;
+}
+
+.records-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.records-table th,
+.records-table td {
+  padding: 12px 14px;
+  text-align: left;
+  border-bottom: 1px solid var(--border-color);
+}
+
+.records-table th {
+  color: var(--text-secondary);
+  font-size: 0.75rem;
+  font-weight: 600;
+}
+
+.records-table tr.late {
+  background: rgba(234, 179, 8, 0.12);
+}
+
+.record-avatar {
+  width: 28px;
+  height: 28px;
+  border-radius: 50%;
+  background: var(--gradient-purple);
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 0.75rem;
+  margin-right: 10px;
+  vertical-align: middle;
+}
+
+.dot {
+  display: inline-block;
+  width: 10px;
+  height: 10px;
+  border-radius: 50%;
+  background: rgba(148, 163, 184, 0.25);
+  margin-right: 4px;
+}
+
+.dot.on {
+  background: var(--accent-purple);
+}
+
+.status-badge {
+  padding: 4px 12px;
+  border-radius: 20px;
+  font-size: 0.8rem;
+}
+
+.status-badge.present {
+  background: rgba(34, 197, 94, 0.2);
+  color: var(--success);
+}
+
+.status-badge.late {
+  background: rgba(234, 179, 8, 0.25);
+  color: var(--warning);
+}
+
+.status-badge.absent {
+  background: rgba(239, 68, 68, 0.2);
+  color: var(--error);
+}
+
+.pagination {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 12px;
+  margin-top: 16px;
+  font-size: 0.875rem;
+  color: var(--text-secondary);
+}
+
+.pagination button {
+  padding: 6px;
+  color: var(--text-secondary);
+}
+
+.summary-stats {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 20px;
+}
+
+.stat-card {
+  padding: 20px;
+  text-align: center;
+}
+
+.stat-card .label {
+  display: block;
+  font-size: 0.8rem;
+  color: var(--text-secondary);
+  margin-bottom: 8px;
+}
+
+.stat-card .value {
+  font-size: 1.75rem;
+  font-weight: 700;
+}
+
+.stat-card .value.green {
+  color: var(--success);
+}
+
+.stat-card .value.yellow {
+  color: var(--warning);
+}
+
+.stat-card .value.red {
+  color: var(--error);
+}
+
+@media (max-width: 1100px) {
+  .two-cards-row {
+    grid-template-columns: 1fr;
+  }
+}
+
+@media (max-width: 900px) {
+  .summary-stats {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
